# Optimize Query Performance in Vet Clinic Database