--- a/schema.sql
+++ b/schema.sql
@@ -1,18 +1,18 @@
 CREATE TABLE owners(
-  id INT GENERATED ALWAYS AS IDENTITY,
+  id SERIAL PRIMARY KEY,
   full_name VARCHAR(20),
   age INT,
   PRIMARY KEY(id)
 );
 
 CREATE TABLE species(
-  id INT GENERATED ALWAYS AS IDENTITY,
+  id SERIAL PRIMARY KEY,
   name VARCHAR(20),
   PRIMARY KEY(id)
 );
 
 CREATE TABLE animals(
-  id INT GENERATED ALWAYS AS IDENTITY,
+  id SERIAL PRIMARY KEY,
   name VARCHAR(20),
   date_of_birth DATE,
   escape_attempts INT,
@@ -24,7 +24,7 @@ CREATE TABLE animals(
 );
 
 CREATE TABLE vets(
-  id INT GENERATED ALWAYS AS IDENTITY,
+  id SERIAL PRIMARY KEY,
   name VARCHAR(20),
   age INT,
   date_of_graduation DATE,
@@ -32,16 +32,20 @@ CREATE TABLE vets(
 );
 
 CREATE TABLE specializations(
-  id INT GENERATED ALWAYS AS IDENTITY,
+  id SERIAL PRIMARY KEY,
   species_id INT REFERENCES species(id),
   vet_id INT REFERENCES vets(id),
   PRIMARY KEY(id)
 );
 
 CREATE TABLE visits(
-  id INT GENERATED ALWAYS AS IDENTITY,
+  id SERIAL PRIMARY KEY,
   animal_id INT REFERENCES animals(id),
   vet_id INT REFERENCES vets(id),
   date_of_visit DATE,
   PRIMARY KEY(id)
 );
+
+CREATE INDEX idx_animal_id ON visits (animal_id);
+CREATE INDEX idx_visits_vet_id ON visits (vet_id);
+CREATE INDEX idx_owners_email ON owners (email);