… screenshots/first decrease.png
… screenshots/first screenshot.png
… screenshots/second decrease.png
… screenshots/second screenshot.png
… screenshots/third decrease.png
… screenshots/third screenshot.png