### Overview

This pull request aims to optimize the performance of specific queries in the Vet Clinic Database, as outlined in the project requirements. The focus was on reducing the execution time for queries that took significantly long to execute (more than 1000ms).

### Changes Made

- Added an index on the `animal_id` column in the `visits` table to optimize `SELECT COUNT(*) FROM visits where animal_id = 4;`.
- Added an index on the `vet_id` column in the `visits` table to optimize `SELECT * FROM visits where vet_id = 2;`.
- Added an index on the `email` column in the `owners` table to optimize `SELECT * FROM owners where email = '[owner_18327@mail.com](mailto:owner_18327@mail.com)';`.

### Performance Improvements

The `EXPLAIN ANALYZE` screenshots below show the execution time before and after the changes:

1. **Query: `SELECT COUNT(*) FROM visits where animal_id = 4;`**
    - **Before**: Execution time was 3864.083 ms.
    ![Before and After Screenshot for Query 1](https://github.com/CVILLA09/vet-clinic2/blob/feature/database-performance-audit/screenshots/first%20screenshot.png?raw=true)

    - **After**: Execution time was reduced to 105.308 ms.
    ![Before and After Screenshot for Query 1](https://github.com/CVILLA09/vet-clinic2/blob/feature/database-performance-audit/screenshots/first%20decrease.png?raw=true)
    
2. **Query: `SELECT * FROM visits where vet_id = 2;`**
    - **Before**: Execution time was 1452.323 ms.
    ![Before and After Screenshot for Query 2](https://github.com/CVILLA09/vet-clinic2/blob/feature/database-performance-audit/screenshots/second%20screenshot.png?raw=true)
    
    - **After**: Execution time was reduced to 901.097 ms.
    ![Before and After Screenshot for Query 2](https://github.com/CVILLA09/vet-clinic2/blob/feature/database-performance-audit/screenshots/second%20decrease.png?raw=true)

3. **Query: `SELECT * FROM owners where email = '[owner_18327@mail.com](mailto:owner_18327@mail.com)';`**
    - **Before**: Execution time was 1134.305 ms.
    ![Before and After Screenshot for Query 3](https://github.com/CVILLA09/vet-clinic2/blob/feature/database-performance-audit/screenshots/third%20screenshot.png?raw=true)
    
    - **After**: Execution time was reduced to 0.137 ms.
    ![Before and After Screenshot for Query 3](https://github.com/CVILLA09/vet-clinic2/blob/feature/database-performance-audit/screenshots/third%20decrease.png?raw=true)

### SQL Files

- Updated `schema.sql` to include the newly added indexes.
- No changes were made to the `data.sql` file as the structure of the database remains the same.

#### Conclusion

This pull request successfully optimizes the performance of the specified queries, making the database more efficient for read operations.